ci: compliance: split CMakeStyle into compliance/ directory
Move the CMakeStyle checker out of the monolithic check_compliance.py
and into scripts/ci/compliance/cmake_style.py, using the auto-loading
infrastructure added in previous commit.
